normalize plugin
