+++ packages/plugin-normalize/README.md
@@ -1,0 +1,3 @@
+# @jimp/plugin-normalize
+
+normalize an image.
